Fix: サーバーシャットダウン時、ユーザーの通信を終了されない不具合を修正

+++ C+lan/C+lan/Time.h
@@ -18,7 +18,7 @@ namespace Ctlan
 				// ----- variables / 変数 ----- //
 				/// <summary>
 				///	デルタタイム	</summary>
-				static double DeltaTime;
+				static float DeltaTime;
 		};
 	}
 }
